Rename files and add new file for API tests in the 'tests/api' directory.
